requirements.txt
pandas==1.5.3
streamlit==1.22.0
streamlit-option-menu==0.3.2
requests==2.28.2
Pillow==9.5.0
openpyxl==3.1.2
XlsxWriter==3.1.2